refactor(http/message): refactoring to use a headers class for header handling
Refs #6.0.x
