+++ src/Http/Message/AbstractMessage.php
@@ -22,17 +22,12 @@ use Phalcon\Http\Message\Exception\InvalidArgumentException;
 use Phalcon\Http\Message\Interfaces\MessageInterface;
 use Phalcon\Http\Message\Interfaces\ResponseStatusCodeInterface;
 use Phalcon\Http\Message\Interfaces\StreamInterface;
-use Phalcon\Http\Message\Interfaces\UriInterface;
-use Phalcon\Support\Collection;
 use Phalcon\Support\Collection\CollectionInterface;
 
 use function array_merge;
 use function implode;
-use function is_array;
-use function is_numeric;
 use function is_resource;
 use function is_string;
-use function preg_match;
 
 /**
  * Message methods
@@ -47,9 +42,9 @@ abstract class AbstractMessage extends AbstractCommon implements MessageInterfac
     protected StreamInterface $body;
 
     /**
-     * @var CollectionInterface
+     * @var Headers
      */
-    protected CollectionInterface $headers;
+    protected Headers $headers;
 
     /**
      * Retrieves the HTTP protocol version as a string.
@@ -60,17 +55,6 @@ abstract class AbstractMessage extends AbstractCommon implements MessageInterfac
      * @var string
      */
     protected string $protocolVersion = "1.1";
-
-    /**
-     * Retrieves the URI instance.
-     *
-     * This method MUST return a UriInterface instance.
-     *
-     * @see http://tools.ietf.org/html/rfc3986#section-4.3
-     *
-     * @var UriInterface
-     */
-    protected UriInterface $uri;
 
     /**
      * Return the body of the stream
@@ -103,7 +87,7 @@ abstract class AbstractMessage extends AbstractCommon implements MessageInterfac
     /**
      * Retrieves a comma-separated string of the values for a single header.
      *
-     * This method returns all of the header values of the given
+     * This method returns all the header values of the given
      * case-insensitive header name as a string concatenated together using
      * a comma.
      *
@@ -162,16 +146,6 @@ abstract class AbstractMessage extends AbstractCommon implements MessageInterfac
     }
 
     /**
-     * Returns the Uri
-     *
-     * @return UriInterface
-     */
-    public function getUri(): UriInterface
-    {
-        return $this->uri;
-    }
-
-    /**
      * Checks if a header exists by the given case-insensitive name.
      *
      * @param string $name
@@ -202,11 +176,11 @@ abstract class AbstractMessage extends AbstractCommon implements MessageInterfac
      */
     public function withAddedHeader(string $name, $value): MessageInterface
     {
-        $this->checkHeaderName($name);
+        $this->headers->checkHeaderName($name);
 
         $headers  = clone $this->headers;
         $existing = $headers->get($name, []);
-        $value    = $this->getHeaderValue($value);
+        $value    = $this->headers->getHeaderValue($value);
         $value    = array_merge($existing, $value);
 
         $headers->set($name, $value);
@@ -256,10 +230,10 @@ abstract class AbstractMessage extends AbstractCommon implements MessageInterfac
      */
     public function withHeader(string $name, $value): MessageInterface
     {
-        $this->checkHeaderName($name);
+        $this->headers->checkHeaderName($name);
 
         $headers = clone $this->headers;
-        $value   = $this->getHeaderValue($value);
+        $value   = $headers->getHeaderValue($value);
 
         $headers->set($name, $value);
 
@@ -311,191 +285,6 @@ abstract class AbstractMessage extends AbstractCommon implements MessageInterfac
     }
 
     /**
-     * Ensure Host is the first header.
-     *
-     * @see: http://tools.ietf.org/html/rfc7230#section-5.4
-     *
-     * @param CollectionInterface $collection
-     *
-     * @return CollectionInterface
-     */
-    final protected function checkHeaderHost(
-        CollectionInterface $collection
-    ): CollectionInterface {
-        if (
-            true === $collection->has("host") &&
-            true !== empty($this->uri) &&
-            "" !== $this->uri->getHost()
-        ) {
-            $host      = $this->getUriHost($this->uri);
-
-            $collection->remove("host");
-
-            $data   = $collection->toArray();
-            $header = ["Host" => [$host]] + (array) $data;
-
-            $collection->clear();
-            $collection->init($header);
-        }
-
-        return $collection;
-    }
-
-    /**
-     * Check the name of the header. Throw exception if not valid
-     *
-     * @see http://tools.ietf.org/html/rfc7230#section-3.2
-     *
-     * @param string $name
-     *
-     * @return void
-     */
-    final protected function checkHeaderName(string $name): void
-    {
-        if (!preg_match("/^[a-zA-Z\d'`#$%&*+.^_|~!-]+$/", $name)) {
-            throw new InvalidArgumentException(
-                "Invalid header name " . $name
-            );
-        }
-    }
-
-    /**
-     * Validates a header value
-     *
-     * Most HTTP header field values are defined using common syntax
-     * components (token, quoted-string, and comment) separated by
-     * whitespace or specific delimiting characters.  Delimiters are chosen
-     * from the set of US-ASCII visual characters not allowed in a token
-     * (DQUOTE and '(),/:;<=>?@[\]{}').
-     *
-     *     token          = 1*tchar
-     *
-     *     tchar          = '!' / '#' / '$' / '%' / '&' / ''' / '*'
-     *                    / '+' / '-' / '.' / '^' / '_' / '`' / '|' / '~'
-     *                    / DIGIT / ALPHA
-     *                    ; any VCHAR, except delimiters
-     *
-     * A string of text is parsed as a single value if it is quoted using
-     * double-quote marks.
-     *
-     *     quoted-string  = DQUOTE *( qdtext / quoted-pair ) DQUOTE
-     *     qdtext         = HTAB / SP /%x21 / %x23-5B / %x5D-7E / obs-text
-     *     obs-text       = %x80-FF
-     *
-     * Comments can be included in some HTTP header fields by surrounding
-     * the comment text with parentheses.  Comments are only allowed in
-     * fields containing 'comment' as part of their field value definition.
-     *
-     *     comment        = '(' *( ctext / quoted-pair / comment ) ')'
-     *     ctext          = HTAB / SP / %x21-27 / %x2A-5B / %x5D-7E / obs-text
-     *
-     * The backslash octet ('\') can be used as a single-octet quoting
-     * mechanism within quoted-string and comment constructs.  Recipients
-     * that process the value of a quoted-string MUST handle a quoted-pair
-     * as if it were replaced by the octet following the backslash.
-     *
-     *     quoted-pair    = '\' ( HTAB / SP / VCHAR / obs-text )
-     *
-     * A sender SHOULD NOT generate a quoted-pair in a quoted-string except
-     * where necessary to quote DQUOTE and backslash octets occurring within
-     * that string.  A sender SHOULD NOT generate a quoted-pair in a comment
-     * except where necessary to quote parentheses ['(' and ')'] and
-     * backslash octets occurring within that comment.
-     *
-     * @see https://tools.ietf.org/html/rfc7230#section-3.2.6
-     *
-     * @param mixed $value
-     *
-     * @return void
-     */
-    final protected function checkHeaderValue($value): void
-    {
-        if (true !== is_string($value) && true !== is_numeric($value)) {
-            throw new InvalidArgumentException("Invalid header value");
-        }
-
-        $value = (string) $value;
-
-        if (
-            preg_match("#(?:(?:(?<!\r)\n)|(?:\r(?!\n))|(?:\r\n(?![ \t])))#", $value) ||
-            preg_match("/[^\x09\x0a\x0d\x20-\x7E\x80-\xFE]/", $value)
-        ) {
-            throw new InvalidArgumentException("Invalid header value");
-        }
-    }
-
-    /**
-     * Returns the header values checked for validity
-     *
-     * @param mixed $values
-     *
-     * @return array
-     */
-    final protected function getHeaderValue($values): array
-    {
-        $valueArray = $values;
-        if (true !== is_array($values)) {
-            $valueArray = [$values];
-        }
-
-        if (true === empty($valueArray)) {
-            throw new InvalidArgumentException(
-                "Invalid header value: must be a string or " .
-                "array of strings; cannot be an empty array"
-            );
-        }
-
-        $valueData = [];
-        foreach ($valueArray as $value) {
-            $this->checkHeaderValue($value);
-
-            $valueData[] = (string) $value;
-        }
-
-        return $valueData;
-    }
-
-    /**
-     * Return the host and if applicable the port
-     *
-     * @param UriInterface $uri
-     *
-     * @return string
-     */
-    final protected function getUriHost(UriInterface $uri): string
-    {
-        $host = $uri->getHost();
-
-        if (null !== $uri->getPort()) {
-            $host .= ":" . $uri->getPort();
-        }
-
-        return $host;
-    }
-
-    /**
-     * Populates the header collection
-     *
-     * @param array $headers
-     *
-     * @return CollectionInterface
-     */
-    final protected function populateHeaderCollection(array $headers): CollectionInterface
-    {
-        $collection = new Collection();
-        foreach ($headers as $name => $value) {
-            $this->checkHeaderName($name);
-
-            $name  = (string) $name;
-            $value = $this->getHeaderValue($value);
-
-            $collection->set($name, $value);
-        }
-
-        return $collection;
-    }
-
-    /**
      * Set a valid stream
      *
      * @param StreamInterface|resource|string $body
@@ -518,32 +307,6 @@ abstract class AbstractMessage extends AbstractCommon implements MessageInterfac
         }
 
         return new Stream($body, $mode);
-    }
-
-    /**
-     * Sets the headers
-     *
-     * @param mixed $headers
-     *
-     * @return CollectionInterface
-     */
-    final protected function processHeaders($headers): CollectionInterface
-    {
-        if (true === is_array($headers)) {
-            $collection = $this->populateHeaderCollection($headers);
-            $collection = $this->checkHeaderHost($collection);
-        } else {
-            if (!($headers instanceof CollectionInterface)) {
-                throw new InvalidArgumentException(
-                    "Headers needs to be either an array or an instance "
-                    . "implementing Phalcon\\Support\\CollectionInterface"
-                );
-            }
-
-            $collection = $headers;
-        }
-
-        return $collection;
     }
 
     /**
